# web/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AparcApp API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/cotxes/enter": {
            "post": {
                "tags": [
                    "Cotxes"
                ],
                "summary": "Register car entry to parking",
                "operationId": "da050df0f80f5c6165332a757d91d413",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "parking_id"
                                ],
                                "properties": {
                                    "parking_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Car entered successfully"
                    },
                    "404": {
                        "description": "Parking not found"
                    },
                    "400": {
                        "description": "Parking is full"
                    }
                }
            }
        },
        "/api/cotxes/exit": {
            "post": {
                "tags": [
                    "Cotxes"
                ],
                "summary": "Register car exit from parking",
                "operationId": "a8c646044912148d21c59c7e2f32556b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "parking_id"
                                ],
                                "properties": {
                                    "parking_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Car exited successfully"
                    },
                    "404": {
                        "description": "Parking not found"
                    },
                    "400": {
                        "description": "Parking is empty"
                    }
                }
            }
        },
        "/api/parking": {
            "get": {
                "tags": [
                    "Parking"
                ],
                "summary": "Get all parking slots",
                "operationId": "b3391eae36b6cead87e9fda88d54a22b",
                "responses": {
                    "200": {
                        "description": "List of all parking slots"
                    },
                    "404": {
                        "description": "No parking slots found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Parking"
                ],
                "summary": "Create new parking slot",
                "operationId": "0d1fe2cf8c0968cca739560fe512159d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "city",
                                    "state",
                                    "zip",
                                    "lat",
                                    "lng",
                                    "capacity",
                                    "occupied"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "zip": {
                                        "type": "string"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "occupied": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Parking slot created successfully"
                    },
                    "400": {
                        "description": "Missing parameters"
                    }
                }
            }
        },
        "/api/parking/{id}": {
            "get": {
                "tags": [
                    "Parking"
                ],
                "summary": "Get specific parking slot",
                "operationId": "a6a0cd57d1951af76c6534cec6635909",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Parking slot ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parking slot details"
                    },
                    "400": {
                        "description": "Missing parameters"
                    }
                }
            },
            "put": {
                "tags": [
                    "Parking"
                ],
                "summary": "Update parking slot",
                "operationId": "48a211465c1510e9cdcadb69da07be98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parking slot ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "city",
                                    "state",
                                    "zip",
                                    "lat",
                                    "lng",
                                    "capacity",
                                    "occupied"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "zip": {
                                        "type": "string"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "occupied": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Parking slot updated successfully"
                    },
                    "400": {
                        "description": "Missing parameters"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Parking"
                ],
                "summary": "Delete parking slot",
                "operationId": "e25ea176ea354b45c5da1b20c3d0a6fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parking slot ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Parking slot deleted successfully"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Parking",
            "description": "Operations about parking"
        },
        {
            "name": "Cotxes",
            "description": "Cotxes management endpoints"
        }
    ]
}